add cam uint32

+++ include/vision/cameraSystem.hpp
@@ -72,7 +72,7 @@ typename std::tuple_element<0, std::tuple<args...> >::type * make_arr(args... ar
 #include <mpl/vision/reprojection.hpp>
 #include <mpl/vision/reconstruction.hpp>
 
-#define CAM(num) (num-1)
+#define CAM(num) (uint32_t)(num-1)
 
 /*****************************************************************************/
 // namespace vision
